Undo darwin user setting
